starting to yank out old game
