+++ src/game/game-init.ts
@@ -26,7 +26,7 @@ import {
 } from "../net/sync.js";
 import { registerPredictSystem } from "../net/predict.js";
 import { registerEventSystems } from "../net/events.js";
-import { initGroundSystem, registerGroundSystems } from "./ground.js";
+import { initRiverTileSystem, registerGroundSystems } from "./river-tile.js";
 import { registerBulletCollisionSystem } from "./bullet-collision.js";
 import { createShip, registerShipSystems, ShipLocalDef } from "./ship.js";
 import { registerBuildBulletsSystem, registerBulletUpdate } from "./bullet.js";
@@ -72,42 +72,12 @@ import { postProcess } from "../render/pipelines/std-post.js";
 import { shadowPipeline } from "../render/pipelines/std-shadow.js";
 import { outlineRender } from "../render/pipelines/std-outline.js";
 
-export const ScoreDef = EM.defineComponent("score", () => {
-  return {
-    maxScore: 0,
-    currentScore: 0,
-  };
-});
-
-function registerScoreSystems(em: EntityManager) {
-  em.addSingletonComponent(ScoreDef);
-
-  em.registerSystem(
-    [ShipLocalDef, PositionDef],
-    [ScoreDef, GameStateDef],
-    (ships, res) => {
-      if (res.gameState.state !== GameState.PLAYING) return;
-      if (ships.length) {
-        const ship = ships.reduce(
-          (p, n) => (n.position[2] > p.position[2] ? n : p),
-          ships[0]
-        );
-        const currentScore = Math.round(ship.position[2] / 10);
-        res.score.maxScore = Math.max(currentScore, res.score.maxScore);
-        res.score.currentScore = currentScore;
-      }
-    },
-    "updateScore"
-  );
-}
-
-export function registerAllSystems(em: EntityManager) {
+export function registerCommonSystems(em: EntityManager) {
   registerNetSystems(em);
   registerInitCanvasSystem(em);
   registerUISystems(em);
   registerDevSystems(em);
   registerGameStateSystems(em);
-  registerScoreSystems(em);
   registerRenderInitSystem(em);
   registerMusicSystems(em);
   registerHandleNetworkEvents(em);
@@ -151,21 +121,6 @@ export function registerAllSystems(em: EntityManager) {
   callInitFns(em);
 }
 
-function registerShipGameUI(em: EntityManager) {
-  em.registerSystem(
-    null,
-    [TextDef, DevConsoleDef],
-    (_, res) => {
-      const avgFPS = 1000 / res.dev.avgFrameTime;
-      const lowerTxt = `Belgus, you are the last hope of the Squindles, keep the gemheart alive! Failure is inevitable. move: WASD, mouse; cannon: e, left-click; fps:${avgFPS.toFixed(
-        1
-      )}`;
-      res.text.lowerText = lowerTxt;
-    },
-    "shipUI"
-  );
-}
-
 function registerRenderViewController(em: EntityManager) {
   em.registerSystem(
     [],
@@ -201,50 +156,5 @@ function registerRenderViewController(em: EntityManager) {
       }
     },
     "renderView"
-  );
-}
-
-export async function initShipGame(em: EntityManager, hosting: boolean) {
-  registerShipGameUI(em);
-  EM.addSingletonComponent(CameraDef);
-  EM.addSingletonComponent(GameStateDef);
-
-  initGroundSystem(em);
-
-  if (hosting) {
-    createShip();
-  }
-
-  // create player once MeDef is present (meaning we've joined, if
-  // we're not the host)
-  em.whenResources([MeDef]).then(() => createPlayer(em));
-
-  const res = await em.whenResources([RendererDef]);
-  res.renderer.pipelines = [
-    shadowPipeline,
-    stdRenderPipeline,
-    outlineRender,
-    postProcess,
-  ];
-}
-
-function debugBoatParts(em: EntityManager) {
-  let once = false;
-  em.registerSystem(
-    [],
-    [AssetsDef],
-    (_, res) => {
-      if (once) return;
-      once = true;
-
-      // TODO(@darzu): this works!
-      // const bigM = res.assets.boat_broken;
-      // for (let i = 0; i < bigM.length; i++) {
-      //   const e = em.newEntity();
-      //   em.ensureComponentOn(e, RenderableConstructDef, bigM[i].mesh);
-      //   em.ensureComponentOn(e, PositionDef, [0, 0, 0]);
-      // }
-    },
-    "debugBoatParts"
   );
 }
